refactor: Create test-common module, consolidate all fakes
New test-common module with all 15 fakes in commonMain:
- One copy of each fake, no duplicates across modules
- Package: com.chriscartland.garage.testcommon
- Depends on domain + data (to see the interfaces)

usecase/commonTest and data/commonTest now depend on test-common
instead of maintaining their own copies.

Deleted: usecase/testfakes/ directory, data/testfakes/ directory
