+++ AndroidGarage/settings.gradle.kts
@@ -45,4 +45,5 @@ include(":domain")
 include(":macrobenchmark")
 include(":presentation-model")
 include(":shared")
+include(":test-common")
 include(":usecase")
