+++ AndroidGarage/test-common/src/commonMain/kotlin/com/chriscartland/garage/testcommon/FakeAuthBridge.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.chriscartland.garage.data.testfakes
+package com.chriscartland.garage.testcommon
 
 import com.chriscartland.garage.data.AuthBridge
 import com.chriscartland.garage.data.AuthUserInfo
