+++ AndroidGarage/test-common/src/commonMain/kotlin/com/chriscartland/garage/testcommon/FakeMessagingBridge.kt
@@ -1,4 +1,4 @@
-package com.chriscartland.garage.data.testfakes
+package com.chriscartland.garage.testcommon
 
 import com.chriscartland.garage.data.MessagingBridge
