+++ AndroidGarage/test-common/src/commonMain/kotlin/com/chriscartland/garage/testcommon/FakeNetworkButtonDataSource.kt
@@ -1,4 +1,4 @@
-package com.chriscartland.garage.data.testfakes
+package com.chriscartland.garage.testcommon
 
 import com.chriscartland.garage.data.NetworkButtonDataSource
 import com.chriscartland.garage.data.NetworkResult
