+++ AndroidGarage/test-common/src/commonMain/kotlin/com/chriscartland/garage/testcommon/FakeNetworkConfigDataSource.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.chriscartland.garage.data.testfakes
+package com.chriscartland.garage.testcommon
 
 import com.chriscartland.garage.data.NetworkConfigDataSource
 import com.chriscartland.garage.data.NetworkResult
