+++ AndroidGarage/test-common/src/commonMain/kotlin/com/chriscartland/garage/testcommon/TestDispatcherProvider.kt
@@ -1,4 +1,4 @@
-package com.chriscartland.garage.usecase.testfakes
+package com.chriscartland.garage.testcommon
 
 import com.chriscartland.garage.domain.coroutines.DispatcherProvider
 import kotlinx.coroutines.CoroutineDispatcher
