+++ AndroidGarage/usecase/src/commonTest/kotlin/com/chriscartland/garage/usecase/DefaultAppLoggerViewModelTest.kt
@@ -1,8 +1,8 @@
 package com.chriscartland.garage.usecase
 
 import com.chriscartland.garage.domain.model.AppLoggerKeys
-import com.chriscartland.garage.usecase.testfakes.FakeAppLoggerRepository
-import com.chriscartland.garage.usecase.testfakes.TestDispatcherProvider
+import com.chriscartland.garage.testcommon.FakeAppLoggerRepository
+import com.chriscartland.garage.testcommon.TestDispatcherProvider
 import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runTest
