+++ AndroidGarage/usecase/src/commonTest/kotlin/com/chriscartland/garage/usecase/RegisterFcmUseCaseTest.kt
@@ -20,8 +20,8 @@ package com.chriscartland.garage.usecase
 import com.chriscartland.garage.domain.model.DoorFcmState
 import com.chriscartland.garage.domain.model.DoorFcmTopic
 import com.chriscartland.garage.domain.model.FcmRegistrationStatus
-import com.chriscartland.garage.usecase.testfakes.FakeDoorFcmRepository
-import com.chriscartland.garage.usecase.testfakes.FakeDoorRepository
+import com.chriscartland.garage.testcommon.FakeDoorFcmRepository
+import com.chriscartland.garage.testcommon.FakeDoorRepository
 import kotlinx.coroutines.test.runTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
